refactor: update icons imports
