+++ src/renderer/components/ReturnButton.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom"
-import { LeftArrowIcon } from "./icons/LeftArrow"
+import { LeftArrow } from "./icons"
 
 interface buttonProps {
     returnTo?: string
@@ -9,7 +9,7 @@ export function ReturnButton({ returnTo = "/" }: buttonProps) {
     return (
         <Link to={returnTo}>
             <div className="left" title="Voltar">
-                <LeftArrowIcon />
+                <LeftArrow />
             </div>
         </Link>
     )
